// frontend/src/types/car.ts
export interface Car {
  id?: number;
  model: string; // This is intentionally 'modal' to match the provided data structure
  year: string; // This is intentionally capitalized to match the provided data structure
  fuelUsage: string;
  driver: string; // This is intentionally capitalized to match the provided data structure
  price: string;
  topSpeed: string;
  image: string;
}